Menu-version-independent aliases
The original set of aliases were built around "old menu" assumptions and, in particular, that the presence of a container-name directory in `~/IOTstack/services` indicated whether the container was in use and, accordingly, whether the aliases should be created.

This was always pretty flimsy and has become more so with "new menu" and the evolving menu system on the "experimental" branch.

The revised approach is to test for the presence of `container_name: «name»` in `docker-compose.yml`.

Because the container subdirectories in `~/IOTstack/services` are going away, all `container_SERVICES` have been removed. That mostly just leaves `container_SHELL` plus some special cases.

`IOTSTACK` as an alias has been replaced with just `I` and augmented with `S`, `T` and `V` for services, templates and volumes, respectively. Each alias changes to the target and runs an `ls`.

> If it seems like the addition of `S` at the same time as removing `container_SERVICES` is a contradiction, the rationale is that `~/IOTstack/services` will likely be required for special cases, such as Node-RED.