Dynamic construction of container shell aliases
All hard-coded shell aliases (eg `NODERED_SHELL`) removed. Script now
iterates running containers and attempts to discover whether each
container supports `bash`, `ash`, `zsh` or `sh` (in that order). The
first match found is used to construct the alias for the container.

Because this is a fairly slow process, the results are cached in:

```
~/.cache/IOTstackAliases/cache
```

which is then sourced by `dot_iotstack_aliases`.

The cache can be rebuilt simply by removing it and re-running
`dot_iotstack_aliases`.

Signed-off-by: Phill Kelley <34226495+Paraphraser@users.noreply.github.com>