fix obscure problem preventing console or VNC access
Consider the following `bash` commands:

```
$ echo $0
-bash

$ CANDIDATES=$(which bash ash zsh sh)
$ echo $CANDIDATES
/usr/bin/bash /usr/bin/sh

$ FIRST=($CANDIDATES)
$ echo $FIRST
/usr/bin/bash
```

In words:

1. Find the paths to the executables for a selection of shells.
2. Take the first member of the responses.

Now consider the same commands executes when `sh` is the shell:

```
$ sh
$ echo $0
sh

$ CANDIDATES=$(which bash ash zsh sh)
$ echo $CANDIDATES
/usr/bin/bash /usr/bin/sh

$ FIRST=($CANDIDATES)
sh: 3: Syntax error: "(" unexpected
$
```

In short, `sh` does not support the `bash` syntax used to extract the
first path from the list of candidates.

IOTstackAliases uses the `FIRST=($CANDIDATES)` syntax. This succeeds
when a user connects via SSH (because `/bin/bash` is the default for
`/etc/passwd`) but fails when a user connects over VNC:

```
/etc/X11/Xsession: nn: /home/pi/.local/IOTstackAliases/dot_iotstack_aliases: Syntax error: "(" unexpected (expecting "fi")
```

In other words, VNC and console connections are using `sh` rather than
`bash`.

Triggering this problem requires an additional factor, which is where
`dot_iotstack_aliases` is sourced.

Consider the following table:

Connection       | .profile     | .bashrc
-----------------|--------------|---------
SSH              | runs         | sourced from .profile
VNC              | runs         | does not run
VNC»terminal     | does not run | runs
Console          | runs         | does not run
Console»terminal | does not run | runs

Because aliases are only useful in terminal sessions, the **correct**
place to source `dot_iotstack_aliases` is `~/.bashrc`. The
IOTstackAliases documentation explains this.

However, if you're:

1. Like me and shovel everything into `.profile` out of long-ingrained
(very bad) habit - not to mention not reading your own documentation; or
2. The victim of my long-ingrained very bad habit because I shovelled
it into the default `.profile` in PiBuilder…

then you can come a right cropper. Sorry about that!

The fix is in two parts. Part one is to replace:

```
FIRST=($CANDIDATES)
```

with:

```
FIRST=$(echo $CANDIDATES | cut -d " " -f 1)
```

That will work for both `bash` and `sh`.

Part two is to alter PiBuilder to use `.profile` and `.bashrc`
**properly**.

This commit also revises the way in which the cache is built. The cache
file will only ever be built when both of the following conditions are
met:

* The cache file does not exist; and
* At least one container is running.

Also fixes a typo in the documentation.

Signed-off-by: Phill Kelley <34226495+Paraphraser@users.noreply.github.com>