handle case where container has no shell
Portainer-CE is an example of a container that has no shell.

> Issue [2768](https://github.com/portainer/portainer/issues/2768)
  includes the claim that this is for "security reasons" but, like many
  such claims, neither explains why nor links to a supportingrationale.

This change defines an alias for any shell-less container which simply
reports `no shell in container`. Defining an alias has the effect of
silencing the "no shell found" report on subsequent logins.

Also took the opportunity to consolidate alias generation and cache
maintenance.

Signed-off-by: Phill Kelley <34226495+Paraphraser@users.noreply.github.com>